+++ src/main/java/com/nhnacademy/frontend/cart/dto/response/BookResponse.java
@@ -24,5 +24,4 @@ public record BookResponse(
         String status,
         int stock
 )
-{
-}
+{}
